# app/static/swagger.yaml
swagger: '2.0'
info:
  version: "0.0.1"
  title: "Mechanic shop API"
  description: API documentation for Mechanic shop
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
...
  /mechanics:
    post:
      tags:
        - "mechanic"
      summary: "Create a new Mechaincs"
      description: "Register a new Mechaincs in the system."
      parameters:
        - in: "body"
          name: "Body"
          description: "Mechaincs object that needs to be added, phone and address are optional"
          required: true
          schema:
            $ref: "#/definitions/MechaincsInput"
      responses:
        201:
          description: "Mechaincs created successfully"
          schema:
            $ref: "#/definitions/MechaincsResponse"
          examples:
            application/json:
              {
                "id": 1,
                "first_name": "John",
                "last_name": "Doe",
                "email": "johndoe@email.com",
                "phone": "1234567890",
                "address": "123 Main St",
                "password": "<encrypted_password>",
                "salary": "50000"
              }
        400:
          description: "Invalid input"
        500:
          description: "Server error"               
  
    get:   
      tags:
        - "mechanic"
      summary: "Get all mechanics"
      description: "Retrieve a list of all mechanics in the system."
      responses:
        200:
          description: "A list of mechanics"
          schema:
            $ref: "#/definitions/mechanicsResponse"
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "first_name": "John",
                  "last_name": "Doe",
                  "email": "testing@email.com",
                  "phone": "1234567890",
                  "address": "123 Main St",
                  "salary": "50000"
                },
                {
                  "id": 2,
                  "first_name": "Jane",
                  "last_name": "Smith",
                  "email": "testing@email.com",
                  "phone": "0987654321",
                  "address": "456 Elm St",
                  "password": "<encrypted_password>",
                  "salary": "60000"  
                }
              ]
        500:
          description: "Server error"

    put: 
      tags:
        - "mechanic"
      summary: "Update a mechanic"
      description: "Update an existing mechanic's information. Requires authentication."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "Body"
          description: "Mechanics object that needs to be updated, phone and address are optional"
          required: true
          schema:
            $ref: "#/definitions/MechaincsInput"
      responses:
        200:
          description: "mechanic updated successfully"
          schema:
            $ref: "#/definitions/MechaincResponse"
          examples:
            application/json:
              {
                "id": 1,
                "first_name": "John",
                "last_name": "Doe",
                "email": "testing@email.com",
                "phone": "1234567890",
                "address": "123 Main St",
                "password": "<encrypted_password>",
                "salary": "50000"
              }
        400:
          description: "Invalid input"
        401:
          description: "Unauthorized - Invalid or missing token"
        404:
          description: "Mechanic not found"
        500:
          description: "Server error"

    delete:
      tags:
        - "mechanic"
      summary: "Delete a mechanic"
      description: "Delete an existing mechanic from the system. Requires authentication."
      security:
        - bearerAuth: []
      responses:
        200:
          schema:
            type: object
            properties:
              message:
                type: string
          description: "mechanic deleted successfully"
          examples:
            application/json:
              {
                "message": "Sorry to see you go, mechanic name!"
              }
        400:
          description: "Invalid input"
        401:
          description: "Unauthorized - Invalid or missing token"
        404:
          description: "Mechanic not found"
        500:
          description: "Server error"     
...
  /ticket_mechanics:  
    post:
      tags:
        - "ticket_mechanics"
      summary: "Assign a mechanic to a service ticket"
      description: "Assign a mechanic to an existing service ticket. Requires authentication."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "Body"
          description: "Object containing service_ticket_id and mechanic_id"
          required: true
          schema:
            type: object
            properties:
              service_ticket_id:
                type: integer
                example: 1
              mechanic_id:
                type: integer
                example: 1
      responses:
        200:
          description: "Mechanic assigned to service ticket successfully"
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              {
                "message": "Mechanic assigned to service ticket successfully!"
              }
        400:
          description: "Invalid input"
        401:
          description: "Unauthorized - Invalid or missing token"
        404:
          description: "Service ticket or mechanic not found"
        500:
          description: "Server error"

    get:
      tags:
        - "ticket_mechanics"
      summary: "Get all ticket-mechanic assignments"
      description: "Retrieve a list of all ticket-mechanic assignments in the system. Requires authentication."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "A list of ticket-mechanic assignments"
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                service_ticket_id:
                  type: integer
                mechanic_id:
                  type: integer
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "service_ticket_id": 1,
                  "mechanic_id": 1
                },
                {
                  "id": 2,
                  "service_ticket_id": 2,
                  "mechanic_id": 2
                }
              ]
        401:
          description: "Unauthorized - Invalid or missing token"
        500:
          description: "Server error"
    put:
      tags:
        - "ticket_mechanics"
      summary: "Update a ticket-mechanic assignment"
      description: "Update an existing ticket-mechanic assignment. Requires authentication."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "Body"
          description: "Object containing service_ticket_id and mechanic_id"
          required: true
          schema:
            type: object
            properties:
              service_ticket_id:
                type: integer
                example: 1
              mechanic_id:
                type: integer
                example: 1
      responses:
        200:
          description: "Ticket-mechanic assignment updated successfully"
          schema:
            type: object
            properties:
              message:
                type: string
          examples:
            application/json:
              {
                "message": "Ticket-mechanic assignment updated successfully!"
              }
        400:
          description: "Invalid input"
        401:
          description: "Unauthorized - Invalid or missing token"
        404:
          description: "Ticket-mechanic assignment not found"
        500:
          description: "Server error"
    delete:
      tags:
        - "ticket_mechanics"  
      summary: "Delete a ticket-mechanic assignment"
      description: "Delete an existing ticket-mechanic assignment from the system. Requires authentication."
      security:
        - bearerAuth: []
      responses:
        200:
          schema:
            type: object
            properties:
              message:
                type: string
          description: "Ticket-mechanic assignment deleted successfully"
          examples:
            application/json:
              {
                "message": "Ticket-mechanic assignment deleted successfully!"
              }
        400:
          description: "Invalid input"
        401:

          description: "Unauthorized - Invalid or missing token"
        404:
          description: "Ticket-mechanic assignment not found"
        500:
          description: "Server error"
...
definitions:
...
  MechaincsInput:
    type: object
    properties:
      first_name:
        type: string
        examples:
          - "John"
      last_name:
        type: string
        examples:
          - "Doe"  
      email:
        type: string
        examples:
          - "test@email.com"
      phone:
        type: string
        examples:
          - "1234567890"
      address:
        type: string
        examples:
          - "123 Main St"
      password:
        type: string
        examples:
          - "testing123"
      salary:
        type: string
        examples:
          - "50000"
...